separated responsibility of js file
