+++ public/stylesheets/style.styl
@@ -10,16 +10,18 @@ a.num
 a.mark
 	width: 65px
 	font-size: 50px
-	
-	
+
 .border_radius
 	border-radius: 5px
 	-webkit-border-radius: 5px
 	-moz-border-radius: 5px
 
+.list-group-item.stripe
+	background-color: #F6F6F6
+
 /* canvas */
 .wall_camera
-	background-color: #FFF
+	background-image: none;
 .wall_wood
 	background-image: url(/image/background/wood.jpg)
 .wall_green_mat
